Small String constexpr pass
A more thorough pass will require C++20 due to a lot of interdependencies

+++ Include/PGE/ResourceManagement/ResourceManager.h
@@ -54,7 +54,7 @@ class ResourceManager {
             deleteResource(view.get());
         }
 
-        void increaseSize(int count);
+        void increaseSize(size_t count);
 };
 
 }
